support fragmented CI TPDU
Otherwise large MMI pages doesn't work (on some CAM modules like
PowerCAM). Will cause en50221_transport.c error:
"Received data with invalid length ..."

from EN 50221:1997 standard:
The first byte of the header is the Transport
Connection Identifier for that TPDU fragment. The
second byte contains a More/Last indicator in its most
significant bit. If the bit is set to '1'
then at least one more TPDU fragment follows,
and if the bit is set to '0'
then it indicates this is the last
(or only) fragment of the TPDU for
that Transport Connection.

Signed-off-by: Abylay Ospan <aospan@netup.ru>